Fix treeForVendor in engines (#49)
* Fix treeForVendor in engines

In an app using engines, if the engines use ember-d3, the d3Modules will be correctly set on the host app only, but the treeForVendor hook will still try to use it on all the engines. This commit fixes this by returning an empty treeForVendor in these cases.

* Fix incorrect usage of this.ui inside of nested function

The dependenciesForPackage method has a different context, leading to this.ui not working.