Use unminified versions of d3 packages.
Debugging through minified code is hard. I noticed that in all of the packages that I was using the `main` pointed at a `.min.js` file but there was also a non-minified version in the same folder.

This change attempts to detect that situation, and use the unminified version if it is present.