Remove import existsSync
Appears to be a relic of some old code. It was causing me errors in my builds as I believe ember and various other packages no longer attempt to install it. Was going to add package to the package.json file however I don't believe it's actually being used by this package any more.